requirements.txt
CrewAI
langchain
openai
python-dotenv